feat(gui): Pygame live viewer — train / play / drive modes
Phases 4-8 of the live viewer (design+plan in docs/superpowers/):
- src/gui render layer (imports ONLY src.sdk): env_view (walls/robot/trail/
  coverage/lidar), curve_view (live reward curve), hud (stats + controls +
  no-checkpoint badge), input_map (key->command / arrow-drive), app (orchestrator:
  update/draw/handle_command). All unit-tested headlessly (SDL dummy driver).
- scripts/play.py: the window loop (argparse map/seed/checkpoint), guarded under
  __main__ so the import-smoke test opens nothing.
- config.gui block (window/fps/speed/trail/demo_checkpoint); assets/demo_policy.pt
  (committed ~1.1 MB) so PLAY works on a fresh clone, else a graceful fallback badge.
- Architecture test extended: scripts may import src.gui; src/gui must import only
  src.sdk (no env/ddpg/services) — keeps business logic behind the SDK.
- §10 now in-scope: docs/UX.md rewritten (3 modes, controls, all 10 Nielsen
  heuristics, accessibility) with per-mode screenshots in assets/screenshots/
  (generated headlessly by scripts/capture_screenshots.py and visually verified);
  README gains a GUI quickstart; audit §10 N/A -> ✅.

Gates: ruff + format + size clean, 290 tests, 97.84% coverage.

+++ src/gui/transform.py
@@ -26,8 +26,8 @@ class View:
         """Map world (x, y) in metres to integer screen pixels (y flipped)."""
         px = self.off_x + (x - self.xmin) * self.scale
         py = self.off_y + (self.span_y - (y - self.ymin)) * self.scale
-        return (int(round(px)), int(round(py)))
+        return (round(px), round(py))
 
     def scale_len(self, metres: float) -> int:
         """Scale a length in metres to a pixel length."""
-        return int(round(metres * self.scale))
+        return round(metres * self.scale)
